summary included
Summary: The purpose of this chart is to show the cumulative amount of total reported COVID-19 cases in the state of Washington by racial groups. Due to the chart's nature being cumulative, a bar chart accurately captures the total value (cases) of the different categories (racial groups). What is notable about this chart is the overwhelming amount of “Unknown” race cases there are that impact the data and alters the information drawn from the set. It is also important to note the large number of cases recorded as “White” and “Hispanic/Latino” compared to the other groups considered. These two observations are incredibly insightful in terms of the disparities seen in recording cases by race, leading to a more narrowed and misleading perception of how the pandemic is affecting specific communities. For example, the Black community and the American Indian community have prominently been heavily impacted and harmed by COVID-19. Still, the data seems to paint a different tale. On the other hand, the number surrounding Hispanic/Latino populations seem to align with what we know in terms of how heavily the virus has ravaged migrant workers in more rural, agricultural parts of Washington.